feat(app): rebuild design system tokens and primitives
Monochrome gray ramp on near-black with rust accent, strict 4px spacing
grid, device-anchored breakpoints, sans UI + mono data fonts. One Button
with variants; single input set; layout primitives (Page, Section, Stack,
Grid); PageHeader. Replaces buttons.tsx/wrappers.tsx grab-bags.
