--- app/src/components/CapTableDashboard.tsx
@@ -1,5 +1,0 @@
-/**
- * Public re-export — implementation lives in ./cap-table.
- */
-export { CapTableDashboard } from "./cap-table";
-export type { CapTableDashboardProps } from "./cap-table";
