+++ app/src/components/global-style.ts
@@ -1,6 +1,16 @@
 import { createGlobalStyle } from "styled-components";
 
 const GlobalStyle = createGlobalStyle`
+	*,
+	*::before,
+	*::after {
+		box-sizing: border-box;
+	}
+
+	:root {
+		color-scheme: dark;
+	}
+
 	html {
 		height: 100%;
 		-webkit-font-smoothing: antialiased;
@@ -14,11 +24,10 @@ const GlobalStyle = createGlobalStyle`
 		min-height: 100%;
 		margin: 0;
 		padding: 0;
-		font-family: "IBM Plex Mono", ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;
+		font-family: ${({ theme }) => theme.fonts.sans};
 		font-weight: ${({ theme }) => theme.fontWeights.normal};
 		font-size: 15px;
 		line-height: ${({ theme }) => theme.lineHeights.P};
-		letter-spacing: -0.01em;
 		color: ${({ theme }) => theme.colors.text};
 		background: ${({ theme }) => theme.colors.background};
 		overflow-y: scroll;
@@ -33,32 +42,22 @@ const GlobalStyle = createGlobalStyle`
 		display: flex;
 	}
 
-	*,
-	*::before,
-	*::after {
-		box-sizing: border-box;
-	}
-
 	::selection {
-		background: ${({ theme }) => theme.colors.main};
-		color: ${({ theme }) => theme.colors.inverse};
+		background: ${({ theme }) => theme.colors.accent};
+		color: ${({ theme }) => theme.colors.onAccent};
 	}
 
 	a {
-		color: ${({ theme }) => theme.colors.main};
+		color: ${({ theme }) => theme.colors.accent};
 		font-weight: ${({ theme }) => theme.fontWeights.medium};
 		text-decoration: none;
 		transition: color ${({ theme }) => theme.transitions.default},
-			opacity ${({ theme }) => theme.transitions.default},
 			text-decoration-color ${({ theme }) => theme.transitions.default};
 
-		/* Never flip body links to white — keeps contrast on dark surfaces */
 		&:hover {
-			color: ${({ theme }) => theme.colors.main};
 			text-decoration: underline;
 			text-decoration-thickness: 1px;
 			text-underline-offset: 0.2em;
-			opacity: 0.92;
 		}
 	}
 
@@ -75,7 +74,16 @@ const GlobalStyle = createGlobalStyle`
 		box-shadow: ${({ theme }) => theme.shadows.focus};
 	}
 
-	/* Subtle custom scrollbars */
+	/* Data is mono — numbers, addresses, hashes, code */
+	code {
+		font-family: ${({ theme }) => theme.fonts.mono};
+		font-size: 0.9em;
+		padding: 0.1em 0.35em;
+		background: ${({ theme }) => theme.colors.surface};
+		color: ${({ theme }) => theme.colors.text};
+	}
+
+	/* Minimal monochrome scrollbars */
 	* {
 		scrollbar-width: thin;
 		scrollbar-color: ${({ theme }) => theme.colors.borderStrong} transparent;
@@ -90,22 +98,21 @@ const GlobalStyle = createGlobalStyle`
 	}
 	::-webkit-scrollbar-thumb {
 		background: ${({ theme }) => theme.colors.borderStrong};
-		border-radius: ${({ theme }) => theme.radii.pill};
 		border: 2px solid transparent;
 		background-clip: padding-box;
 	}
 
-	code {
-		font-family: inherit;
-		font-size: 0.9em;
-		padding: 0.1em 0.35em;
-		border-radius: ${({ theme }) => theme.radii.sm};
-		background: ${({ theme }) => theme.colors.input};
-		color: ${({ theme }) => theme.colors.main};
-	}
-
 	input, select, textarea {
 		color-scheme: dark;
+	}
+
+	@media (prefers-reduced-motion: reduce) {
+		*,
+		*::before,
+		*::after {
+			animation-duration: 0.001ms !important;
+			transition-duration: 0.001ms !important;
+		}
 	}
 `;
