+++ app/src/components/shell/navConfig.ts
