+++ app/src/components/typography.tsx
@@ -1,77 +1,64 @@
 import styled from "styled-components";
 
+/**
+ * Typography — sans for UI copy, mono for data.
+ * Text blocks cap at theme.maxWidths.text (~680px) for legibility.
+ */
+
 const H1 = styled.h1`
-	margin: 0 0 ${({ theme }) => theme.spacing.xl} 0;
-	max-width: ${({ theme }) => theme.maxWidths.h1};
+	margin: 0;
+	max-width: ${({ theme }) => theme.maxWidths.text};
 	font-size: ${({ theme }) => theme.fontSizes.H1};
 	line-height: ${({ theme }) => theme.lineHeights.H1};
-	font-weight: ${({ theme }) => theme.fontWeights.semibold};
-	letter-spacing: -0.04em;
-	color: ${({ theme }) => theme.colors.text};
-`;
-
-const H2 = styled.h2`
-	margin: 0 0 ${({ theme }) => theme.spacing.lg} 0;
-	max-width: ${({ theme }) => theme.maxWidths.h1};
-	font-size: ${({ theme }) => theme.fontSizes.H2};
-	line-height: ${({ theme }) => theme.lineHeights.H2};
 	font-weight: ${({ theme }) => theme.fontWeights.semibold};
 	letter-spacing: -0.03em;
 	color: ${({ theme }) => theme.colors.text};
 `;
 
-const H3 = styled.h3`
-	margin: 0 0 ${({ theme }) => theme.spacing.md} 0;
-	max-width: 40rem;
-	font-size: ${({ theme }) => theme.fontSizes.H3};
-	line-height: ${({ theme }) => theme.lineHeights.H3};
+const H2 = styled.h2`
+	margin: 0;
+	max-width: ${({ theme }) => theme.maxWidths.text};
+	font-size: ${({ theme }) => theme.fontSizes.H2};
+	line-height: ${({ theme }) => theme.lineHeights.H2};
 	font-weight: ${({ theme }) => theme.fontWeights.semibold};
 	letter-spacing: -0.02em;
 	color: ${({ theme }) => theme.colors.text};
 `;
 
-const Blockquote = styled.blockquote`
-	margin: 0 0 ${({ theme }) => theme.spacing.xl} 0;
-	padding: ${({ theme }) => theme.spacing.md} ${({ theme }) => theme.spacing.lg};
-	max-width: 46rem;
-	border-left: 2px solid ${({ theme }) => theme.colors.main};
-	background: ${({ theme }) => theme.colors.accentMuted};
-	border-radius: 0 ${({ theme }) => theme.radii.md} ${({ theme }) => theme.radii.md} 0;
-	font-size: ${({ theme }) => theme.fontSizes.large};
-	line-height: 1.55;
+const H3 = styled.h3`
+	margin: 0;
+	max-width: ${({ theme }) => theme.maxWidths.text};
+	font-size: ${({ theme }) => theme.fontSizes.H3};
+	line-height: ${({ theme }) => theme.lineHeights.H3};
+	font-weight: ${({ theme }) => theme.fontWeights.semibold};
+	letter-spacing: -0.01em;
 	color: ${({ theme }) => theme.colors.text};
 `;
 
 const P = styled.p`
-	margin: 0 0 ${({ theme }) => theme.spacing.md} 0;
-	max-width: 40rem;
+	margin: 0;
+	max-width: ${({ theme }) => theme.maxWidths.text};
 	font-size: ${({ theme }) => theme.fontSizes.medium};
 	line-height: ${({ theme }) => theme.lineHeights.P};
 	font-weight: ${({ theme }) => theme.fontWeights.normal};
-	color: ${({ theme }) => theme.colors.muted};
-	letter-spacing: -0.01em;
+	color: ${({ theme }) => theme.colors.textMuted};
 
-	@media only screen and (max-width: ${({ theme }) => theme.breakpoints.mobile}) {
+	@media (max-width: ${({ theme }) => theme.breakpoints.phone}) {
 		font-size: ${({ theme }) => theme.fontSizes.baseline};
 	}
 `;
 
+/** Uppercase micro-label for fields and table headers. */
 const Label = styled.label`
 	display: inline-block;
 	font-size: ${({ theme }) => theme.fontSizes.xs};
 	font-weight: ${({ theme }) => theme.fontWeights.semibold};
 	letter-spacing: 0.08em;
 	text-transform: uppercase;
-	color: ${({ theme }) => theme.colors.subtle};
+	color: ${({ theme }) => theme.colors.textSubtle};
 `;
 
-const OrderedList = styled.ol`
-	margin: ${({ theme }) => theme.spacing.lg} 0;
-	padding-left: ${({ theme }) => theme.spacing.xl};
-	font-size: ${({ theme }) => theme.fontSizes.medium};
-	color: ${({ theme }) => theme.colors.muted};
-`;
-
+/** Accent kicker above a heading. */
 const Eyebrow = styled.span`
 	display: inline-flex;
 	align-items: center;
@@ -80,7 +67,23 @@ const Eyebrow = styled.span`
 	font-weight: ${({ theme }) => theme.fontWeights.semibold};
 	letter-spacing: 0.12em;
 	text-transform: uppercase;
-	color: ${({ theme }) => theme.colors.main};
+	color: ${({ theme }) => theme.colors.accent};
 `;
 
-export { H1, H2, H3, Blockquote, P, Label, OrderedList, Eyebrow };
+/** Secondary explanatory copy. */
+const MutedText = styled.p`
+	margin: 0;
+	max-width: ${({ theme }) => theme.maxWidths.text};
+	font-size: ${({ theme }) => theme.fontSizes.small};
+	line-height: ${({ theme }) => theme.lineHeights.P};
+	color: ${({ theme }) => theme.colors.textMuted};
+`;
+
+/** Data span — numbers, addresses, tx hashes. Always mono, tabular. */
+const Mono = styled.span`
+	font-family: ${({ theme }) => theme.fonts.mono};
+	font-variant-numeric: tabular-nums;
+	letter-spacing: 0;
+`;
+
+export { H1, H2, H3, P, Label, Eyebrow, MutedText, Mono };
